Run: Don’t hang on large outputs from a script
This patch fixes a bug that would cause `marathon run` to hang when the
script being run produced a large amount of output. The problem was that
the pipe Marathon uses to run the script would get filled, causing the
pipe to hang.

The solution is to read from the pipe before waiting for it to exit.