PackageManager: Resolve name of package from Package.swift instead of URL
Previously, Marathon was using the URL of an added package to resolve its name
(for example if the URL ended in “Files.git” the package name would be assumed
to be “Files”). While simple, this technique was not very robust.

The new way of doing it, with this change, is to clone any remote repository
added, and then parse the name from the package’s Package.swift file. If it’s
a local repository we just read the file directly from the filesystem.