Inject & overload ‘print’ to avoid unwanted output in tests
Currently, when you run the tests you will see output from Marathon mixed
in between the output from `XCTest`. That’s ugly - so let’s fix it! This
change makes `Marathon` take a `printer`, which is essentially just a
printing function to use, which is then used to overload `print` whenever
output is printed on the command line.