Support verbose output with '--verbose' flag
This change introduces a '--verbose’ flag that can be passed to any command,
which will make all underlying command line commands printed as output.

This is really useful when debugging problems or finding the underlying cause
for some errors.

In doing this, the usage of printing and printers have also been reworked,
printing is now done through a central object with 3 levels of importance:

- output
- progress
- verbose