ScriptManager: Add support for inline dependency resolution
This change adds support for specifying dependencies inline, instead of
having to use a Marathonfile. A comment can now simply be added next to
an `import` statement, specifying the URL that the dependency can be found
at, and Marathon will automatically add that dependency, just like when a
Marathonfile is used.

The implementation is a bit naive, as it simply splits the script file based
on new lines and iterates through them (although an optimization has been made
to stop as soon as a non-import alphanumeric character is found). A more
optimized solution would probably be to stream the file and keep reading bytes
until all imports have been parsed.