PackageManager: Don’t re-add already added packages with differently cased URL
This patch fixes a bug that would cause a package to be re-added to Marathon when the
given URL had a different casing than the one already added. So, if first Files is added
using `https://github.com/JohnSundell/Files`, it would then be re-added if the URL
`https://github.com/johnsundell/files` was given.