Run: Enable running a remote script using a URL, repository or GitHub path
This change makes `run` able to run a remote script from a URL, a git
repository URL or a GitHub path. For example, in order to run Test Drive
users can now use either of these three methods to run it directly, without
having to install it first:

- `marathon run https://github.com/johnsundell/testdrive.git`
- `marathon run https://raw.githubusercontent.com/JohnSundell/TestDrive/master/Sources/TestDrive.swift`
- `marathon run johnsundell/testdrive`

The latter is the big improvement, that enables running remote scripts
using a very concise syntax.