Add test assuring that all tests run on Linux
This change introduces a `testAllTestsRunOnLinux()` test that will look at
the test source code and make sure that all tests have been added to the
`allTests` array that is used to run tests on Linux.

The only tests that are excluded are the ones that contain `MacOS` in their
name (that are verifying things like using `xcrun` and Xcode project generation).