Improve error message when a script was not found
Previously, running “edit” or “run” with a missing script path would cause
Marathon to start looking for a remote script. With this patch, remote scripts
are only looked up for the “run” and “install” commands (since you can’t create
or edit a remote script), and the error messages have been improved to include
better actionable information if a script is missing.

Resolves https://github.com/JohnSundell/Marathon/issues/94