Make Swift Package Manager cache packages when using Xcode 9
This patch prevents the Swift Package Manager from re-cloning and re-building
all dependency packages for each script, thanks to some symlinking. We already
use this technique for other files, but the Swift toolchain bundled with Xcode 9
includes two new files that need to be symlinked in order for SPM not to invalidate
its state.