Support scripting using Swift 4
With this change Marathon supports writing scripts in Swift 4. If the user
is using a Swift 4 toolchain, the language version will automatically be set
to 4.0. Earlier toolchains will still use Swift 3.x.

The Swift 4 version of the Swift Package Manager requires a few more parameters
to be specified in the package manifest, so this change mainly focuses on adding
those, as well as adding a module folder (which is also required in Swift 4).

Backwards compatibility with Swift 3.1+ is still maintained.