Support running/installing remote Swift packages as scripts
This change makes it possible to have Marathon run and install remote
executable Swift packages as scripts. Since what Marathon does under
the hood is basically just to generate Swift packages for script files,
all that needed to be done to support this new feature is to check if
a downloaded repository is indeed a package, and then just use it directly.

To test that this works, Marathon remotely installs and runs itself, from
its repo (so meta!)

Resolves https://github.com/JohnSundell/Marathon/issues/120