Support Swift 5 scripting
This change makes it possible to use Marathon with Swift 5 and
the latest Xcode beta. The only change required to Marathon itself
was to correct the code that compared Swift versions (which currently
was hard-wired to compare against `4.2`).

The other changes required were all in the tests, which were making
assumptions about Swift Package Manager output that no longer are
true.

This is a 100% backward compatible change.